# scanner: fix UTF-8 and control character handling