The scanner operates on bytes, but to do proper UTF-8 validation, we
must transition some sections of the scanner to rune validation.

Benchmarks show no difference in memory allocations but somewhat random
CPU results on my laptop:

```
name                               old time/op    new time/op    delta
UnmarshalDataset/config-3            20.2ms ± 1%    21.8ms ± 1%   +7.99%  (p=0.008 n=5+5)
UnmarshalDataset/canada-3            89.6ms ±11%    83.6ms ± 3%     ~     (p=0.222 n=5+5)
UnmarshalDataset/citm_catalog-3      33.0ms ± 2%    27.9ms ± 6%  -15.41%  (p=0.008 n=5+5)
UnmarshalDataset/twitter-3           13.3ms ± 3%    11.6ms ± 1%  -12.68%  (p=0.008 n=5+5)
UnmarshalDataset/code-3               114ms ± 6%      98ms ± 4%  -13.65%  (p=0.008 n=5+5)
UnmarshalDataset/example-3            212µs ± 2%     208µs ± 1%     ~     (p=0.056 n=5+5)
Unmarshal/SimpleDocument/struct-3     727ns ± 1%     708ns ± 4%     ~     (p=0.222 n=5+5)
Unmarshal/SimpleDocument/map-3       1.09µs ± 3%    1.04µs ± 3%   -5.09%  (p=0.008 n=5+5)
Unmarshal/ReferenceFile/struct-3     46.6µs ± 3%    56.1µs ± 2%  +20.35%  (p=0.008 n=5+5)
Unmarshal/ReferenceFile/map-3        78.2µs ± 1%    86.3µs ± 5%  +10.40%  (p=0.008 n=5+5)
Unmarshal/HugoFrontMatter-3          15.5µs ± 1%    16.4µs ± 5%   +5.75%  (p=0.008 n=5+5)
Marshal/SimpleDocument/struct-3       564ns ± 3%     553ns ± 3%     ~     (p=0.222 n=5+5)
Marshal/SimpleDocument/map-3          878ns ± 3%     868ns ± 2%     ~     (p=0.690 n=5+5)
Marshal/ReferenceFile/struct-3       46.3µs ± 1%    46.6µs ± 2%     ~     (p=0.690 n=5+5)
Marshal/ReferenceFile/map-3          67.1µs ± 2%    67.2µs ± 7%     ~     (p=0.841 n=5+5)
Marshal/HugoFrontMatter-3            12.3µs ± 3%    12.5µs ±11%     ~     (p=1.000 n=5+5)

name                               old speed      new speed      delta
UnmarshalDataset/config-3          52.0MB/s ± 1%  48.1MB/s ± 1%   -7.41%  (p=0.008 n=5+5)
UnmarshalDataset/canada-3          24.7MB/s ±10%  26.4MB/s ± 3%     ~     (p=0.222 n=5+5)
UnmarshalDataset/citm_catalog-3    16.9MB/s ± 2%  20.0MB/s ± 6%  +18.36%  (p=0.008 n=5+5)
UnmarshalDataset/twitter-3         33.3MB/s ± 3%  38.1MB/s ± 1%  +14.48%  (p=0.008 n=5+5)
UnmarshalDataset/code-3            23.7MB/s ± 6%  27.4MB/s ± 4%  +15.75%  (p=0.008 n=5+5)
UnmarshalDataset/example-3         38.2MB/s ± 2%  39.0MB/s ± 1%     ~     (p=0.056 n=5+5)
Unmarshal/SimpleDocument/struct-3  15.1MB/s ± 1%  15.5MB/s ± 4%     ~     (p=0.222 n=5+5)
Unmarshal/SimpleDocument/map-3     10.1MB/s ± 3%  10.6MB/s ± 3%   +5.36%  (p=0.008 n=5+5)
Unmarshal/ReferenceFile/struct-3    113MB/s ± 3%    93MB/s ± 2%  -16.92%  (p=0.008 n=5+5)
Unmarshal/ReferenceFile/map-3      67.0MB/s ± 1%  60.7MB/s ± 5%   -9.37%  (p=0.008 n=5+5)
Unmarshal/HugoFrontMatter-3        35.3MB/s ± 1%  33.4MB/s ± 5%   -5.39%  (p=0.008 n=5+5)
Marshal/SimpleDocument/struct-3    21.3MB/s ± 3%  21.7MB/s ± 3%     ~     (p=0.167 n=5+5)
Marshal/SimpleDocument/map-3       13.7MB/s ± 3%  13.8MB/s ± 2%     ~     (p=0.690 n=5+5)
Marshal/ReferenceFile/struct-3     44.1MB/s ± 1%  43.8MB/s ± 2%     ~     (p=0.595 n=5+5)
Marshal/ReferenceFile/map-3        29.7MB/s ± 2%  29.7MB/s ± 7%     ~     (p=0.841 n=5+5)
Marshal/HugoFrontMatter-3          42.3MB/s ± 3%  41.9MB/s ±10%     ~     (p=1.000 n=5+5)
```

Updates: #613