@moorereason 2021-10-05
Fixes at least the following test pattern:

```
go test -tags testsuite -run='TestTOMLTest_.*(Codepoint|Control|Unicode)' ./...
```