@pelletier 2021-10-05
Benchmark on my desktop are definitely slower on this version:

```
name                                old time/op    new time/op    delta
UnmarshalDataset/config-32            23.0ms ± 0%    24.3ms ± 1%   +5.78%  (p=0.000 n=10+9)
UnmarshalDataset/canada-32            80.6ms ± 1%    79.9ms ± 1%   -0.81%  (p=0.001 n=10+9)
UnmarshalDataset/citm_catalog-32      26.2ms ± 1%    26.4ms ± 1%   +0.89%  (p=0.000 n=9+10)
UnmarshalDataset/twitter-32           10.1ms ± 1%    10.7ms ± 1%   +5.50%  (p=0.000 n=10+10)
UnmarshalDataset/code-32              98.2ms ± 0%   100.4ms ± 1%   +2.16%  (p=0.000 n=10+10)
UnmarshalDataset/example-32            169µs ± 0%     179µs ± 1%   +5.90%  (p=0.000 n=10+9)
Unmarshal/SimpleDocument/struct-32     566ns ± 1%     588ns ± 0%   +3.79%  (p=0.000 n=10+10)
Unmarshal/SimpleDocument/map-32        861ns ± 1%     873ns ± 1%   +1.45%  (p=0.000 n=10+10)
Unmarshal/ReferenceFile/struct-32     38.0µs ± 0%    49.3µs ± 0%  +29.71%  (p=0.000 n=9+10)
Unmarshal/ReferenceFile/map-32        66.7µs ± 0%    77.2µs ± 1%  +15.86%  (p=0.000 n=9+10)
Unmarshal/HugoFrontMatter-32          13.9µs ± 0%    15.0µs ± 0%   +7.74%  (p=0.000 n=10+10)
```

To be expected though: it's doing more work.

I think the rune decoding could be skipped: we only care about validation. Could it be done in a second pass? First pass to find the end of the string, then second pass to validate that the bytes in the string form a valid UTF8 sequence. That would open up the use of faster validation methods like [`utf8.Valid`](https://pkg.go.dev/unicode/utf8#Valid) or [a SIMD version](https://arxiv.org/pdf/2010.03090.pdf).