@moorereason 2021-10-11
Life has become very busy for me recently.  I'm not sure when I'll be able to rework this PR.  Do with it as you please.