@pelletier 2021-10-14
@moorereason totally understand! I always appreciate all the time and effort you've put in this project, thank you!
I'll work on this patch when I have time.